feat: [조안나] flutter app bar naming modify
